refactor(registration): move hackathon-participation question to step 1
Relocates the "Have you participated in other events" select from
Step 3 to the bottom of Step 1's "Additional information" section,
next to the founder/ecosystem flags. The field, schema entry, and
i18n keys (reg.step2.hackathon.*) are unchanged — only the rendering
location moves. Step 3 sheds the now-orphan Select imports and the
options constant; Step 1 picks them up.
